Filter with select on search operation
