+++ app/operations/perilune/tasks/find_operation.rb
@@ -4,7 +4,7 @@ module Perilune
   module Tasks
     class FindOperation < LedgerSync::Domains::Operation::Find
       def serializer_for(*)
-        Perilune::Tasks::Serializer.new
+        Perilune::Tasks::Show::Serializer.new
       end
     end
   end
